training time is now calculated from activities with breaks
